requirements_dashboard.txt
streamlit>=1.35.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pandas>=2.0.0
plotly>=5.18.0
python-dotenv>=1.0.0
requests>=2.31.0